Add files via upload
# Multiple Search Program
This C++ program generates and displays multiples of a user-defined number. The program allows the user to input a number, and it generates the first 10 multiples of that number. It also allows the user to search for a specific multiple in the list of generated multiples.

## Features

Input a number for which multiples will be generated.

- Outputs the first 10 multiples of the given number.
- Allows the user to search for a specific multiple in the list.
- Displays whether the searched multiple is found and its position, or if it is not a multiple.